Add size_t

+++ ch8/pass-reference.cpp
@@ -2,14 +2,14 @@
 #include <vector>
 
 void init(std::vector<double>&x){
-  for(unsigned int i=0;i<x.size();++i){
+  for(size_t i=0;i<x.size();++i){
     x[i]=1.*i;
   }
 }
 
 void printv(const std::vector<double>& x){
   std::cout<<"{";
-  for(unsigned int i=0;i<x.size();++i){
+  for(size_t i=0;i<x.size();++i){
     std::cout<<x[i];
     if(i<x.size()-1) std::cout<<", ";
   }
